JPA fix

+++ API/src/main/java/com/Springzada/API/repository/FazendaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FazendaRepository extends JpaRepository<FazendaModel, Integer>{
+public interface FazendaRepository extends JpaRepository<FazendaModel, Long>{
 
     FazendaModel findByNome(String nome);
